requirements.txt
pytesseract==0.3.13
python-dotenv==1.0.1
nipype==1.8.6
pillow==10.4.0
aiofiles==24.1.0
uvicorn==0.30.6
fastapi==0.112.2
PyMuPDF
python-multipart